+++ app/src/main/java/com/babytracker/designsystem/components/topbar/TopBar.kt
@@ -39,8 +39,10 @@ fun AppTopBar(
     titleWeight: FontWeight = TopBarDefaults.titleWeight(),
     backIconSize: Dp = TopBarDefaults.backIconSize(),
     actions: @Composable RowScope.() -> Unit = {},
+    modifier: Modifier = Modifier,
 ) {
     CenterAlignedTopAppBar(
+        modifier = modifier,
         title = {
             Text(
                 title,